python_acpi/requirements.txt
numpy>=1.26.0
Pillow>=9.0.0
matplotlib>=3.8.0
scipy>=1.11.0
